requirements.txt

pandas
numpy
scipy
h5py
dill
pillow
. e -
